refactor(supernova): refactored all non filtered alerts from zustand to RQ cache
