+++ apps/supernova/src/lib/createSilencesSlice.jsx
@@ -172,44 +172,6 @@ const createSilencesSlice = (set, get, options) => ({
           })
         })
       },
-
-      /*
-      Find all silences (also expired ones) in items that matches all labels (key&value) from the alert but omit the labels that are excluded (excludedLabels)
-      */
-      getSilencesForAlert: (alert) => {
-        if (!alert) return
-
-        const alertLabels = alert?.labels || {}
-
-        // collect all silences
-        let silences = [...get().silences.items]
-
-        // collect all excluded Labels
-        const excludedLabels = get().silences.excludedLabels || []
-        const enrichedLabels = get().alerts.enrichedLabels || []
-        // combine the arrays containing the labels that shouldn't be used for matching into one for easier checking
-        const labelsExcludedForMatching = [...excludedLabels, ...enrichedLabels]
-
-        // Find all silences where all silence.labels are a subset of alerts.labels (excluding the excluded excludedLabels)
-        return silences.filter((silence) => {
-          const silenceMatchers = silence?.matchers || []
-
-          // Check if all non-excluded labels from the silence are included in the alert
-          return silenceMatchers.every((silenceLabel) => {
-            if (labelsExcludedForMatching.includes(silenceLabel.name)) return true
-            const alertLabelValue = alertLabels[silenceLabel.name]
-
-            // If the label is not a regex, check if the values are equal
-            if (!silenceLabel.isRegex) {
-              return silenceLabel.value === alertLabelValue
-            }
-
-            // SilenceLabel.isRegex is true so we need to check if the alertLabelValue matches the regex
-            const regex = new RegExp(silenceLabel.value)
-            return regex.test(alertLabelValue)
-          })
-        })
-      },
       /*
         Returns the silence (including the local ones) with the latest expiration time for an alert. Useful to display when the alert will be active again.
       */
